refactor(common/*): move common components
