+++ app/components/common/DeletePanel/constants.ts
