+++ app/components/common/DeletePanel/index.tsx
