+++ app/features/Pairing/Panel/FilterItemOptionButton/index.tsx
@@ -1,4 +1,4 @@
-import Button from '~/components/Button';
+import Button from '~/components/common/Button';
 
 export type Props = {
   isPressed: boolean;
