+++ app/features/Pairing/Panel/index.tsx
@@ -1,6 +1,6 @@
 import Filter from './Filter';
 import Setting from './Setting';
-import Button from '~/components/Button';
+import Button from '~/components/common/Button';
 import { usePairContext } from '~/features/Pairing/context';
 import { SUBMIT_BUTTON } from './constants';
