+++ app/features/Pairing/Portal/index.tsx
@@ -1,6 +1,6 @@
 import { usePairContext } from '../context';
 import { HeaderPortal } from '~/components/Layout/header';
-import IconButton from '~/components/IconButton';
+import IconButton from '~/components/common/IconButton';
 import { Fragment } from 'react';
 
 export default function Portal() {
