+++ app/features/hospital/cards/index.tsx
@@ -1,4 +1,4 @@
-import Icon from '~/components/Icon';
+import Icon from '~/components/common/Icon';
 import { Hospital } from './types';
 import { City } from '~/features/hospital/constants';
 import { LABEL_OF_OPEN_DURING_COVID } from './constants';
